+++ src/main/java/sellFan/dao/iterface/IBillDAO.java
@@ -6,5 +6,5 @@ import java.math.BigInteger;
 
 public interface IBillDAO extends GenericDAO<Bill>{
     public int createBill(Bill bill);
-    public void updateTotal(int billId, BigInteger total);
+    public void updateTotal(int billId, int total);
 }
